Suport Sugar versions as a proxy for the Kernel
This provides a Sugar\Manager that will make use of a Sugar\Finder\Finder to
find a sugar based on a given path.
Sugar\Manager will give you a proxy so that you can mock all Sugar calls based
on their versions. This will allow us to work faster providing just the changes
of each version and keep the same functionality on the commands (being version
agnostic as much as possible).
Keep in mind that you can still check for the Sugar Version to make special
cases on your commands.